PersonService & AddressService & AllergyTest unit tests implementations

+++ src/main/java/com/openclassrooms/safety_net/service/AllergyService.java
@@ -3,7 +3,9 @@ package com.openclassrooms.safety_net.service;
 import com.openclassrooms.safety_net.model.Allergy;
 import com.openclassrooms.safety_net.repository.AllergyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -24,12 +26,20 @@ public class AllergyService {
 		return StreamSupport.stream(allergies.spliterator(), false).toList();
 	}
 
-	public void deleteAllergy (Allergy allergy) {
-		allergyRepository.delete(allergy);
+	public void deleteAllergy (Allergy allergy) throws ResponseStatusException {
+		if (allergyRepository.existsById(allergy.getId())) {
+			allergyRepository.delete(allergy);
+		} else {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Allergy %s not found and can't be deleted.".formatted(allergy.getName()));
+		}
 	}
 
 	public void deleteAllergiesWithoutMedicalRecordLink () {
 		Iterable<Allergy> allergies = getAllergiesOnlyWithNoLinks();
-		allergies.forEach(this::deleteAllergy);
+		try {
+			allergies.forEach(this::deleteAllergy);
+		} catch (ResponseStatusException e) {
+			e.printStackTrace();
+		}
 	}
 }
